Add DocketEntryDocumentList component and tests
Jira ticket: CAMS-262
